Backend Bulk Changes

+++ splitwise-backend/src/main/java/com/splitwise/backend/dto/BalanceResponse.java
@@ -2,12 +2,12 @@ package com.splitwise.backend.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
 public class BalanceResponse {
     private String fromUserId;
     private String toUserId;
-    private double amount;
+    private BigDecimal amount;
 }
-
