+++ splitwise-backend/src/main/java/com/splitwise/backend/dto/ExpenseResponse.java
@@ -4,6 +4,7 @@ import com.splitwise.backend.model.SplitType;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 import java.util.Map;
@@ -14,11 +15,11 @@ public class ExpenseResponse {
 
     private String id;
     private String description;
-    private double amount;
+    private BigDecimal amount;
     private SplitType splitType;
     private String paidByUserId;
     private List<String> participantUserIds;
-    private Map<String, Double> splitDetails;
+    private Map<String, BigDecimal> splitDetails;
     private Instant createdAt;
     private Instant updatedAt;
 }
